Export disabled argTypes from .storybook/preview.js
